Rename files for clarity and for general northern spirit code use
