+++ ex2_hal/northern_spirit/equipment_handler/include/northern_file_transfer.h
