+++ ex2_hal/northern_spirit/equipment_handler/source/northern_file_transfer.c
@@ -17,8 +17,7 @@
  * @date 2022-04-12
  */
 
-#include "ft_2u_payload_handler.h"
-
+#include <northern_file_transfer.h>
 #include <string.h>
 
 #include "stdio.h"
